refactor(student): extract ActivityPage custom hooks
